compiler: one file per target and per language's tests; the tables refusal derived from what registered
compiler/builtin.go's target types move to compiler/target_<lang>.go, each
registering itself from init through registerBuiltin(g, tables); the js file
is target_javascript.go because Go reads a _js suffix as a build constraint.
refuseTables names the table-carrying targets from that registration rather
than from a typed list, and its test refuses through a probe target so it
keeps testing when the last tableless built-in gains its backend.

compiler/tables_test.go's per-language tests move to tables<lang>_test.go,
beside the existing tablesjs_test.go, with the helpers only they use; the
shared helpers stay. The CLI's --lang help lists what Targets() returns.
